#  perf: improve href getter performance