Before

```
---------------------------------------------------------------------------------------
Benchmark                             Time             CPU   Iterations UserCounters...
---------------------------------------------------------------------------------------
BasicBench_AdaURL              32524681 ns     32480095 ns           21 GHz=3.22738 cycle/byte=12.0204 cycles/url=1044.08 instructions/byte=48.6195 instructions/cycle=4.04475 instructions/ns=13.054 instructions/url=4.22305k ns/url=323.507 speed=267.49M/s time/byte=3.73846ns time/url=324.72ns url/s=3.07958M/s
BasicBench_AdaURL_just_parse   17782300 ns     17774775 ns           40 GHz=3.22596 cycle/byte=6.52503 cycles/url=566.759 instructions/byte=26.3285 instructions/cycle=4.03501 instructions/ns=13.0168 instructions/url=2.28687k ns/url=175.687 speed=488.788M/s time/byte=2.04588ns time/url=177.703ns url/s=5.62736M/s
```

After

```
---------------------------------------------------------------------------------------
Benchmark                             Time             CPU   Iterations UserCounters...
---------------------------------------------------------------------------------------
BasicBench_AdaURL              30952299 ns     30951273 ns           22 GHz=3.22836 cycle/byte=11.4879 cycles/url=997.83 instructions/byte=47.3469 instructions/cycle=4.12146 instructions/ns=13.3055 instructions/url=4.11252k ns/url=309.083 speed=280.702M/s time/byte=3.56249ns time/url=309.435ns url/s=3.23169M/s
BasicBench_AdaURL_just_parse   17426274 ns     17426275 ns           40 GHz=3.22926 cycle/byte=6.47108 cycles/url=562.073 instructions/byte=26.3955 instructions/cycle=4.07899 instructions/ns=13.1721 instructions/url=2.29269k ns/url=174.056 speed=498.563M/s time/byte=2.00577ns time/url=174.219ns url/s=5.7399M/s
```